+++ lib/open_weather/models/forecast.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 require_relative 'forecast/hourly'
+require_relative 'forecast/thirty_day'
 require_relative 'forecast/forecast'
 require_relative 'forecast/city'